Change extra information and finish README.md
Add real name in extra information and remove the nod to it being HTML. Also
complete the README.md file.

My real name should be there since it is a genuine HTML project that I've
completed and can add to my portfolio, which is why I made the change. Also,
the project being HTML is obvious and is refered to in my README.md,
so I don't need to have it displayed on the website. Since I've finished
most major changes to my project, I wanted to solidify it and put the
cherry on top by writing the README.md, explaining my experiences and
what the premise behind the project is.